+++ package.json
@@ -9,10 +9,12 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
-    "bootstrap": "^5.0.2",
+    "bootstrap": "^5.1.0",
     "firebase": "^8.8.1",
+    "jquery": "^3.6.0",
     "jshint": "^2.13.0",
     "react": "^17.0.2",
+    "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
     "react-firebaseui": "^5.0.2",
     "react-router-dom": "^5.2.0",
