+++ src/components/UserProfile/PhotographerProfile.css
@@ -1,0 +1,8 @@
+.tableHolder {
+  width: 70vw;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 50px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+}
